+++ clients/fides-js/src/components/Overlay.tsx
@@ -1,10 +1,6 @@
 import { h, FunctionComponent, VNode } from "preact";
 import { useEffect, useState, useCallback, useMemo } from "preact/hooks";
-import {
-  FidesOptions,
-  PrivacyExperience,
-  ServingComponent,
-} from "../lib/consent-types";
+import { FidesOptions, PrivacyExperience } from "../lib/consent-types";
 
 import { debugLog, hasActionNeededNotices } from "../lib/consent-utils";
 
@@ -30,6 +26,7 @@ interface Props {
   options: FidesOptions;
   experience: PrivacyExperience;
   cookie: FidesCookie;
+  onOpen: () => void;
   renderBanner: (props: RenderBannerProps) => VNode | null;
   renderModalContent: (props: RenderModalContent) => VNode;
   onVendorPageClick?: () => void;
@@ -39,6 +36,7 @@ const Overlay: FunctionComponent<Props> = ({
   experience,
   options,
   cookie,
+  onOpen,
   renderBanner,
   renderModalContent,
   onVendorPageClick,
@@ -63,9 +61,7 @@ const Overlay: FunctionComponent<Props> = ({
   const handleOpenModal = useCallback(() => {
     if (instance) {
       instance.show();
-      dispatchFidesEvent("FidesUIShown", cookie, options.debug, {
-        servingComponent: ServingComponent.OVERLAY,
-      });
+      onOpen();
     }
   }, [instance, cookie, options.debug]);
 
@@ -121,15 +117,6 @@ const Overlay: FunctionComponent<Props> = ({
       !options.fidesEmbed,
     [experience, options]
   );
-
-  useEffect(() => {
-    const eventCookie = cookie;
-    if (showBanner && bannerIsOpen) {
-      dispatchFidesEvent("FidesUIShown", eventCookie, options.debug, {
-        servingComponent: ServingComponent.BANNER,
-      });
-    }
-  }, [showBanner, cookie, options.debug, bannerIsOpen]);
 
   const handleManagePreferencesClick = (): void => {
     handleOpenModal();